Add weak pointers `GcWeak` and `GcWeakCell`
This is a combination of work by EmperorBale and me.
It adds weak-pointer version of `Gc` and `GcCell`, named
`GcWeak` and `GcWeakCell` respectively. These can be created
by calling `downgrade()` on the corresponding 'strong' struct
(`Gc`/`GcCell`).

You can attempt to upgrade a weak pointer to a strong pointer
by calling `upgrade`. This will succeed as long as the target
object has not yet been freed (it will also fail during `Phase::Sweep`
when the target object is going to be freed, but has not yet been
freed).

To support this, `GcBox` now has two additional flags: `alive`
and `has_weak_ref`. When an allocation with weak pointers is
garbage-collected, we just run the destructor on the `Collect`
value if there are still weak pointers. The `GcBox` itself is
not deallocated, so that the weak pointers are still able to access
the flags. In a later `Phase::Sweep` where all of the weak pointers
are gone, we actually deallocate the memory for the `GcBox`.

I've also added comments to existing code, to make it clear how
weak pointers interact with the existing invariants.